% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jshintr.R
\name{jshintr}
\alias{jshintr}
\title{Lint a Javascript file}
\usage{
jshintr(filepath, width = NULL, height = "auto", elementId = NULL)
}
\arguments{
\item{filepath}{path to a JavaScript file}

\item{width, height}{dimensions (CSS properties)}

\item{elementId}{a HTML id for the widget (usually useless)}
}
\value{
A \code{htmlwidget} object.
}
\description{
HTML widget which runs 'jshint' on the given JavaScript file.
}
\note{
You can also run this widget as a RStudio addin.
}
\examples{
jsfile <- system.file("jsexample.js", package = "jshintr")
cat(readLines(jsfile), sep = "\n")
jshintr(jsfile)
}
